Remove config parameters unused in application code
This removes a pattern that was somewhat confusing where we configured
classes to accept config options as arguments and didn't exercise these
at all.

This approach was likely used for testing but it seems that over time
the tests have moved on. As these config options are static methods it
would be more idiomatic to use mocking rather than injection.

My motivation for removing this was wondering what purpose these
arguments offered before concluding that there didn't seem to be any.