Calls to Rails.logger in a Sidekiq process use the Sidekiq logger
Previously any calls to Rails.logger in a Sidekiq process would use that
class and presumably be written to stderr in production, meaning it's
rather unlikely this information would be seen in the context of jobs.

By setting the Rails logger to Sidekiq logger any calls to Rails.logger
during the Sidekiq process will be wrote to the Sidekiq log. This is
particularly useful for classes that aren't workers that want to log an
event, with the previous approach these would need to consider the
environment they were in to choose a logger, whereas with this approach
it'll write to the correct log depending on your context.

If this approach works without issues in Email Alert API I'll plan to
port it into govuk_sidekiq.