Fix Zeitwerk loading EmailAlertAPI constant
This resolves a problem where the Email Alert API application was not
starting when using the Zeitwerk autoloader. The application wasn't
starting because it expected that the `lib/email_alert_api/config`
directory would define a constant EmailAlertApi::Config and when it
didn't it raised an error.

```
/Users/kevindew/.rbenv/versions/2.6.6/lib/ruby/gems/2.6.0/gems/zeitwerk-2.4.0/lib/zeitwerk/loader.rb:392:in `const_get': uninitialized constant EmailAlertApi::Config (NameError)
```

This was a problem that only affects the production environment because
this is the only environment that has config.eager_load set to true. In
other environments classes are autoloaded when needed.