Add task to switch users to daily digest emails
This adds a new task as part of an experiment to switch immediate
subscribers of specific lists to daily digest emails. For each
subscriber, we send a summary email of the changes made. In order to
track the changes, we use a new reason for start/ending a subscription.

We process each subscriber individually, using the same locking approach
as when a subscriber is manually updating their subscriptions. I decided
to avoid doing a single, bulk change, as it's not clear that a roll back
would be possible in the presence of one or more race conditions.

The task is best effort: if a race condition causes the change to fail
for a particular subscriber, we can always rerun the task.