Remove redundant warning metric
Stop sending a warning latency metric for content changes and messages.
The use of this has been removed in this [puppet change].

As we've removed the warning alert then it makes sense to make the
actual metric more descriptive of what it's actually showing instead of
being concerened on whether it is specific to a critical or warning
state.

We should deploy this at the same time as the [puppet change].

Trello:
https://trello.com/c/BT5x4iAU/457-update-warning-critical-alerts-for-messages-and-content-changes

[puppet change]: https://github.com/alphagov/govuk-puppet/pull/10580