Index digest_run_subscribers by subscriber
Currently it's impossible to query this model for a large number of
subscribers, since these records are not naturally ordered by this
column. This adds an index to facilitate a query we want to run, to
determine the number of digest emails sent to a set of subscribers.

Note that, although we wish to run this query of a time period, an
index on "created_at" has little added benefit, since the records
are naturally ordered by this column as-is.