Remove manual cleanup rake task
It's unclear why this task was added, since it relates to a deeper
problem with the quality of data in the system, that we should look
at fixing at source. This removes the task, since we don't plan to
run it periodically.

Added in: https://github.com/alphagov/email-alert-api/pull/1188