Mark a digest run as processed when all subscribers identified
This sets the processed at value for a digest run when the subscriber
identification is complete. This is an early step in the work to allow
us to retry running a digest should any problems occur. The intention is
that when a DigestInitiatorService runs it will check whether a digest
run is processed or not, if it is not it will try to initiate the
digest. Thus, this field serves as an indication that the DigestRun
finished it's job.