Replace DigestRun completion methods
We had a few methods going on here, `mark_complete!`,
`check_and_mark_complete!` and `has_incomplete_subscribers?` which
seemed a touch abundant. I've replaced them all with a single
`mark_as_completed` method on the model (which is consistent with Email)
and moved the check for incomplete subscribers to the worker that
is performing this operation.