Add a unique index for digest_run_subscribers
This is part of work to make it possible to retry running a failed
digest run. Currently if you try to retry a failed digest run it will
create duplicate instances in digest_run_subscribers and there is
nothing to prevent that.

This change makes it so that PostgreSQL will reject any insertions that
conflict with existing ones.

I benchmarked adding this index on integration:

17:02:54 Migrating to UniqueIndexOnDigestRunSubscribers (20200818150135)
17:02:54 == 20200818150135 UniqueIndexOnDigestRunSubscribers: migrating ================
17:02:54 -- add_index(:digest_run_subscribers, [:digest_run_id, :subscriber_id], {:unique=>true, :algorithm=>:concurrently})
17:04:36    -> 102.6307s
17:04:36 == 20200818150135 UniqueIndexOnDigestRunSubscribers: migrated (102.6309s) =====

I experimented to see if it had an adverse effect on the insert times
for running a DailyDigestInitiatorWorker run. Much to my surprise the
inserts were actually quicker (!!) with the index, both inserted 236,417
records into DigestRunSubscriber.

Before: done: 1291.694 sec
After: done: 918.023 sec