Skip duplicates when importing DigestRunSubscribers
In the previous commit we introduced a unique index for this table in
order to ensure integrity was preserved in this table even if a failed
digest run was retried. This commit takes the next step and changes the
insert configuration to skip duplicate records, this will mean that if we
resume processing a digest run it will try to insert all
DigestRunSubscriber records but will skip any that already exist.