Rename completed_at to processed_at on DigestRun
This renames this column so that this model can be consistent with other
models that also use processed_at to indicate when the worker process
has completed for that model.

Normally the process to rename a column for a Rails app is rather
tricky, with the application raising errors at any points in time where
the code for the app is running a different version. This often means
renaming a column is a complex process of creating a new column and
backfilling it before deleting the old one. In this situation though I
determined that this could be done by a simple rename so long as this
was deployed outside the time that digests are running as that is the
only time this column is accessed. Thus, as long as this is deployed
carefully it, this can go out in a single step.

I tried running this on integration to see if the migration was fast and
it was very quick:

18:31:37 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] Migrating to RenameCompletedAtDigestRunSubscriber (20200818172715)
18:31:37 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] == 20200818172715 RenameCompletedAtDigestRunSubscriber: migrating =============
18:31:37 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] -- rename_column(:digest_run_subscribers, :completed_at, :processed_at)
18:31:37 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] -> 0.0109s
18:31:37 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal] == 20200818172715 RenameCompletedAtDigestRunSubscriber: migrated (0.0110s) ====
18:31:37 *** [err :: ip-10-1-5-9.eu-west-1.compute.internal]