Improve performance of the switch-to-digest task
Previously the task to switch individual subscriptions to a daily
digest had several performance issues:

- Building the full "subscriptions_by_subscriber" hash would lead
to an OOM error for the number of lists we wish to migrate.

- Processing each subscriber took a significant amount of time, to
the extent the time to run the job would be many hours.

This rewrites the task to perform the migration in batches of the
subscribers affected, reducing the memory footprint. It also makes
several changes to reduce the number of DB queries:

- Including related models to avoid individual queries for them
- Rolling all subscription changes into two queries per user

As a result of these changes, the expected job time is now ~1 hour.