Adapt DigestEmailGenerationWorker for allowing retries
This adapts this worker so that it can abort elegantly should it be
called in a situation it be re-queued after completing by checking
whether the item is processed or not.

It also applies a FOR UPDATE database lock on the DigestRunSubscriber
object when this runs, this is to create isolation that will prevent any
concurrent attempts to run this worker - without this if we ended up
with this worker running twice, creating two emails and leaving the app
in a somewhat inconsistent state.

I also refined the tests a little by adding a more specific clause to
creating an email and removing a superfluous one that tests we don't
change digest run completed_at.