Rename non subscription_content uses of the term
I found myself a bit confused by the code that referred to
subscription_content which wasn't referencing the model and instead this
structure we've created of DigestSubscriptionContentQuery::Result
objects. In particular this was confusing in DigestEmailGenerationWorker
where it is also creating SubscriptionContent models.

The approach I've taken is to rename the DigestSubscriptionContentQuery
class to DigestItemsQuery, which seems to work since we only have one
list concept that comprises of a digest. I've then updated the naming
of subscription_content to digest_items.