Fix prematurely marking DigestRuns as complete
As the worker to check if DigestRun objects are complete runs every
minute it can mark a DigestRun as complete before it is created any
DigestRunSubscriber objects and thus mark the digest complete before
this has actually occurred.

This can be resolved by using the new processed_at field to check that a
DigestRun is also processed before making this check.

I started this change hoping that I could actually remove the
completed_at field as it seemed a bit of a smell that we run a worker
every minute of every day for it. However doing so would sacrifice some
of the relevance of the Metrics::DigestRunExporter and thus I have left
it. I'm hoping that in the not too distant future we can see an
opportunity to remove it.