Close subscriptions for FCO/DFID lists
This adds a rake task which identifies lists that are associated with
FCO and DFID organisations and closes any subscriptions to them. It
filters out any lists that are associated with additional organisations.

This task has been put together to reflect that these lists will no
longer receive any emails now that the FCO and DFID organisations are
closed and that these subscriptions have become superfluous.

Once this has task has been run on September 8th (as per the current
schedule) it can be deleted from the code.