Reinstate simplified report for content changes
Previously we had a report for content changes matched against each
subscriber list with its immediate subscribers, as a way to discover
the origin of email volume [1]. Since then, we have begun reporting
email volume by publishing app and document type [2], and have done
migrations to increase the use of daily digest emails, to the extent
that immediate subscriber numbers are no longer meaningful as a way
to estimate volume, and are time consuming to include. This reinstates
a report of just the matched content changes, to support our research
into the behaviour of specific lists.

[1]: https://github.com/alphagov/email-alert-api/pull/1265
[2]: https://github.com/alphagov/email-alert-api/pull/1323