Replace console analytics with rake tasks
Previously our "analytics.md" doc specified an arbitrary set of DB
queries to get insights. We also moved a couple of rake tasks to
clarify that they are not for the purpose of "support" [1]. This
replaces the arbitrary textual suggestions of DB queries with a list
of the supported rake tasks for extracting data about the system.

[1]: https://github.com/alphagov/email-alert-api/pull/1371