Stop sending finished_sending_at & sent to athena
The reason we are stopping sending these is they may change after an
email has been archived, which makes them misleading. For archiving
purposes, it should be enough to know when the email was created, since
this is usually similar to when it was sent.

Also removes an unused let in EmailArchiveQuery.

Trello:
https://trello.com/c/15eVh9Qg/482-stop-relying-on-notify-status-updates-to-archive-delete-emails