Add validation for weekly digests on a Saturday
This is so we can be confident that we won't accidentally send users
a digest email with overlapping contents.

Another way of validating this would be to check two digests with the
same range do not overlap. However, this is more likely to create a
problem where someone unintentionally creates a digest one day later
than normal, which would have missing contents.