Prevent duplicate concurrent digest jobs
This repeats the technique we used for retrying content change and
message processing, such that we can freely enqueue work that exists,
in the knowledge it will be a no-op. Using this technique means we
can avoid more elaborate, manual checks for work-in-progress.

Note that, because the job can be executed without arguments by the
scheduler, we need to apply the same default when checking uniqueness.