Lock digest initiation by range *and* date
Previously we guarded against repeat initiation of daily and weekly
digests, even when these were for different dates. While this would
eventually fix itself due to repeat recovery attempts, and should
also be highly unlikely (a previous digest would have to be broken
for over a day), the locking should be consistent with the arguments.