Replace sidekiq-unique-jobs with non-blocking lock
Previously we added sidekiq-unique-jobs as a way to prevent duplicate
jobs executing concurrently. However, we've since learned the gem is
flawed for the main scenario we're trying to prevent: a worker being
killed due to memory issues [1]. An non-blocking lock [2] can be used
to get the same no-wait-noop effect, similarly to an advisory lock, but
without the need to define a unique namespace [3]. This replaces the
use of the gem with non-blocking locks, which are more robust.

Note this replaces the use of specific models with ApplicationRecord
where the specific model class is not relevant.

Note that, while we could use "SKIP LOCKED" instead of "NOWAIT", this
would still result in a "not found" error later on, due to the record
being invisible to all but the process holding the lock [2]. I think
it's better to use "NOWAIT", so the error we catch (to avoid a failure
/ retry) is explicitly related to locking.

[1]: https://github.com/alphagov/email-alert-api/pull/1412#issuecomment-695999823
[2]: https://www.postgresql.org/docs/9.5/sql-select.html
[3]: https://github.com/alphagov/email-alert-api/pull/1405/commits/e2b85f9e1223d3af386a6c20029962445e6ac0c3