Favour DB locks over sidekiq-unique-jobs
Previously we used a gem to prevent enqueueing or even repeating
ongoing work, in the event that it is recovered prematurely by the
RecoverLostJobsWorker. In order to achieve this, we used a strategy
of "until_executed", which has a flaw that it can prevent legitimate
recovery if a previous attempt dies e.g. due to a restart [1].

This introduces advisory locks around the processing of content
changes and messages, so we can stop using the faulty gem. Note that
other workers using the gem already implement DB-level locks. While
we could switch to a strategy of "until_execut**ing**", to maintain
some of the benefit of the gem, it's confusing to have two de-dup
mechanisms next to each other. DB locks address our main concern.

Note also that advisory locks are not dependent on the model/table
we're working with: DigestRun.with_advisory_lock("id") is the same
as ApplicationRecord.with_advisory_lock("id").

[1]: https://github.com/alphagov/email-alert-api/pull/1412#issuecomment-695999823