DRY-up workers with a base class
Previously we had two kinds of repetitive code in our workers:

- "include Sidekiq::Worker"
- various code to run with an advisory lock

This DRYs-up the code into a new base class. As part of doing this,
we need to standardise the ID structure of advisory locks, which
means this change will need to be deployed during a period when
none of the affected workers have work outstanding. To be sure
of this, we can disable non-critical workers temporarily, and do
the deployment after any digest runs have finished processing.

Summary of worker locks:

- ProcessContentChangeWorker - lock is new (not affected)
- ProcessMessageWorker - lock is new (not affected)
- DailyDigestInitiatorWorker - deploy when completed for the day
- WeeklyDigestInitiatorWorker - deploy during the week
- EmailArchiveWorker - non-critical, disable temporarily
- EmailDeletionWorker - non-critical, disable temporarily
- NullifyDeactivatedSubscribersWorker - non-critical, disable temporarily