Add an endpoint to get a subscriber's details from an account token
For at least our initial experiment, we want to be able to manage
email subscriptions from the account manager.  All email-alert-api
interaction is done in terms of subscriber IDs, rather than addresses,
and the only way to turn an address into a subscriber ID is by sending
a verification email with a magic link in it.

We don't want to have a "double verification email" scenario, where we
ask a user to verify their account and then also to verify we can
manage their subscriptions.  So add this endpoint which:

1. Accepts a JWT over POST containing the fields `email` and `email_verified`
2. Checks that the JWT has a valid signature from the account system
3. Checks that the `email_verified` is true
4. Returns either an error or the subscriber details

With the ID and the address all the other endpoints work.  We can:

- List a user's subscriptions
- Change a user's address
- Unsubscribe a user from one or all of their subscriptions
- Subscribe a user to a list (which sends them a confirmation, but not
  verification, email)
- Update the frequency of a subscription