Set-up rake task for experiment 3 of immediate -> daily
In experiment 3 there is a new dimension to the criteria of how lists
will be converted - proportion - where some lists are only having a
percentage of their subscribers converted from immediate subscriptions
to daily digest. The consequences of this are: input to this task is no
longer a single dimension array and running the task isn't idempotent
(it will convert 20% of however many subscribers there are at the point
of execution).

To implement this I've cannibalised the switch_to_daily_digest rake task
to be solely focused on the experiment. This was done since the input
was no longer simple and reflected that the existing task wasn't
actually used for distinct operations (it is indeed coupled to the
experiment numbering used in the builder class), it does carry the
downside that we'll have more to delete if we temporarily remove the
rake task.

The input has switched from being a text file to a CSV, the simplest way
to display a proportion seemed to be a decimal - this avoided the need
to do percentage calculations.