Schedule the deletion of historical data
Schedule HistoricalDataDeletionWorker to run every day at midday.

The bulk of records to be deleted are digest run subscribers which we
are currently producing at around ~400,000 per day, these on average are
created within 30 minutes in the morning of each digest run.

The combined amount of the rest deletions per day from the other models
comes to under 10k per day.

We could delete in batches for digest run subscribers but wasn't sure
that was necessary as in integration it only took around 13 seconds to
delete 400k digest run subscribers which is fairly quick.