Remove code to initiate experiment
This code was used to start experiment 3 of switching immediate
subscriptions to digests. This is being removed since this exact same
task won't be run again.

There is a good chance we'll want to resurrect this for further
experiments however the precedent on previous experiments has been to
delete the code and then resurrect it when needed.

The CSV file and the report task will remain until work monitoring the
experiment has completed.