Stop creating delivery attempt models when sending email
The delivery attempt model is largely redundant now that we only try to
send an email once and that it now basically just maintains a log of
Sidekiq attempts which is rather unconventional.

This changes the reference that is sent to Notify to be the id of the
email being sent. As we don't try send emails multiple times this is
unique per email.

I was able to move the ProviderCommunicationFailureError exception into
DeliveryRequestService as we no longer have the requirement that this
should result in a persisted delivery attempt.

A side effect of this change is that we now increment the rate limit
prior to making an attempt to send an email. This has the effect that we
will always increment the rate limit whether we are successful or not.
It also means that it will be incremented regardless of whether a
request to Notify was made or not in staging/integration - which means
they will need to have a rate limit that is the same as prod to be able
to process emails at a similar pace.

The metrics we store about time until first delivery attempt no
longer make much sense now we only send an email to Notify once. I've
removed the one for time for email creation to email first delivery
attempt as we don't make use of that figure. For the content change one I've
created a new name that we can migrate to and kept the old one running
for continuity so that the dashboard can be switched over without data
loss.

[1]: https://grafana.blue.production.govuk.digital/dashboard/file/email_alert_api_technical.json?refresh=1m&orgId=1