Remove rate limit lever env vars
In the previous commit we started incrementing the rate limiter
regardless of whether we're sending an email or not. This has an impact
in integration and staging where we try send a large volume of emails
but most are aborted (due to not having an address on an allowed list).

In order for these environments to perform acceptably they should use
the same rate limit as production. We send so few emails in integration
and staging that is very unlikely we'll hit any actual service rate
limits so we don't explicitly need a lower rate limit configured for
these environments. Were we to wish to do so it would make sense to move
the rate limit code from the worker to the provider so they can be
appropriately coupled.

[1]: https://github.com/alphagov/govuk-puppet/pull/10787