Move rate_limiter out of Services module
There's no reason that this class needs to be available across the whole
application and always be the same instance.

To avoid initialising a fresh instance of Redis for each worker we can
use the one that Sidekiq already has available, we can also make use of
that being under a namespace so we no don't need to prefix it with
email-alert-api.