Migrate status-updates to be email id focussed
In the proceeding commits we changed the email sending to no longer use
the delivery attempt model. This continues this change by applying this
to the status update controller.

A somewhat significant change is that we decided we don't have a huge
use for storing the outcome of the email sending in the database. This
data has no bearing on the application and is purely useful for audits.
Since emails have a short life (7 days) it seems more prudent to add
this auditing in when we have a practical need, rather than a presumed
need. A side effect of this is that the 404 behaviour changes as we no
longer need to look-up emails.

Without storing data on each request this code can be simpler as we just
need to focus on handling permanent-failure statuses and any that
are cause for alarm (such as technical-failure). Because of this I
moved this code from a dedicated service class to be handled within the
controller.

Should someone wish to know if a particular email is sent or not this
data is now available in the log and it can also be looked up on
using Notify's UI.