Rename DeliveryRequest{Worker,Service}
The names for these were always rather confusing (after all have you
heard any one describe sending an email as "requesting the delivery of
an email?") and this seems somewhat compounded by the deprecation and
coming removal of the DeliveryAttempt model.

Using the term delivery is a somewhat loaded concept in Email Alert API
terminology. In recent times we've switched to describing the job of
Email Alert API as _sending_ an email to Notify, it is a later stage in
the pipeline that the email is actually delivered and whether an email
is delivered more nuanced than sending it (this is because there can be
problems at the client side that result in an email not being
delivered).

I've gone with a more common terminology here and used the phrasing of
send email.