Stop using delivery attempt id to look up Notify emails
Now that we are dropping the DeliveryAttempt concept emails in Notify
need to be looked up by the email id rather than the delivery attempt
id. This switch allows this code to be a reasonable amount simpler.

There is a caveat that this task won't work for existing emails sent
using the current DeliveryAttempt models however as this task is used
only infrequently it seems quite reasonable to go without it and just
warn 2nd line support about it.