Update sending email queue names to reflect sending
These were wrapped up in "delivery" terminology which we are moving
away. We have adjusted our focus to be more on the sending of email
than the management of email delivery and are adapting terminology to
reflect this.

This change keeps the old queues so they're items can be exhausted in
tandem with this. This will also require updating to documentation and
alerts that alert based on queue sizes.