Allow all emails to be archivable and deleteable
This change allows all emails to be archivable which allows them also to
be deleted.

Previously emails were only archived if they did not have a status of
pending. This was needed previously as we were sending extra data to
[Athena] which was only accessible after a Notify callback for each
email. As we've changed our [email delivery success strategy] to not
rely on status updates to mark an email as sent, we've also removed
callback fields from being archived. This means we can now change our
archiving policy to include pending emails as soon as they're created.

This is part of the cleanup work so we don't end up with lingering
emails and subscriptions contents in the system which will never be
deleted or recovered.

This commit also removes the scopes of 'deleteable' and 'archiveable'
as they each only had one filter so seemed unecessary to keep.

Trello:
https://trello.com/c/tfu9zXuy/534-delete-unused-data-after-a-year

[Athena]: https://github.com/alphagov/email-alert-api/commit/090df0e2221a3597f2d1d6f36a0b8d2dac7d27d7
[email delivery success strategy]: https://github.com/alphagov/email-alert-api/pull/1402