Some clean up of the BrexitSubscribersReport
This class had something rather peculiar going on with the arguments. It
omitted the asterisk when forwarding the arguments which meant that an
array was passed into the initialiser. The only reason this code
actually worked with this is because of the to_s call on `Date.parse`.
For example if you called this with `Date.today` then the call to
`Date.parse` was "[Wed, 21 Oct 2020]" which actually worked.

To resolve this I've switched this class around to operate on a date
object rather than a string or array as before.

I also noticed that the error message when this gets an incorrect
argument didn't abort the task so I've switched that to a raise.