Catch Net::ReadTimeout errors from Notify
We saw one of these come through to our error handling today. I had
thought we had covered timeouts previously with checking for
Net::OpenTimeout and didn't realise it had a sibling.

I did consider whether to catch the superclass of these, Timeout::Error,
but decided that this wasn't the best approach as Timeout::Error seems
unrelated to Net communications and I felt that it would be strange to
either a) not test the exceptions we expect or b) test against different
exceptions than the code.