Stop having fancy output for rake tasks
Previous we had an inline task method (so difficult to test) that
formatted the output for a couple of the tasks. Turns out it's
actually _harder_ to read the tabulated output because because it
ends up wrapping on narrow windows. This removes the formatting, so
we just output simple data structures, which are perfectly readable.