Update nullify worker and rename
We can work out if a user should be nullified without the need of the
`deactivated_at` field which is soon to be removed. If they have no active
subscriptions and their most recent is over 28 days old then they should
be nullified.

We should also nullify subscribers which have no subscriptions which are
older than 28 days. We don't expect their to be subscribers without
subscriptions but as seen in later commits this has happened before. We
should remove these regularly so we're complying with GDPR and not
retaining users data even if there is a system bug.