Switch to only migrated non-migrated subscribers
Previously we could assume that all subscribers for a list had not
been migrated. Since we have now "touched" all lists with previous
iterations of this task, we need to identify the subscribers who we
have yet to migrate; these are the subscribers who do not have a
migrated subscription (active or ended) associated with a given list.
This updates the migration task to only migrate subscribers that
have not been previously migrated, irrespective of how many times
they may have switched back and forth since.