Create subscriber & subscription inside transaction
This is another avenue where subscribers could be left with no
subscriptions.

We should create the subscriber and subscription in the same transaction
when a user first creates their subscription. This is so we aren't left
in a state where a subscriber exists with no subscription if an error
occurs in the subscription creation process.

This is relevant to a later commit where we're going to cleanup
subscribers with no subscriptions instead of relying on the
`deactivated_at` field.