Update UnsubscribeAllService
Two things:

* Deletes UnsubscribeService and inlines the code to unsubscribe one
  subscription in the places it's used.

* Updates UnsubscribeAllService. The code originally looped through the
  active subscriptions updating one by one, a simpler approach which
  mitigates the need for transaction block is to update all the
  subscriptions for a subscriber in one using `update_all`.

We also need to record metrics when unsubscribing so the code for this
has been added. To do this we've updated the `unsubscribed` method to
use statsd's `count` instead of `increment` method which allows us to
update the metric by more than one at a time.

Finally in this commit, as a bit of a tidy up, we stopped using `.last`
when only searching for one record in the support tasks. Using `find_by`
instead of `where` was more fitting.