Backfill ended subscriptions updated_at
In [this PR] a bug was introduced which stopped updating the
`updated_at` when all of a users subscriptions were ended. This had a
knock on effect of breaking a [report].

Also updates all the other instances where the ended_at time does not
match updated_at.

Disabling the wrapping transaction means if the migration fails for
whatever reason, you could re-run it and pick up where the migration
left off as there are quite a few records to update (629,854).

[this PR]: https://github.com/alphagov/email-alert-api/pull/1462#discussion_r519839598
[report]: https://github.com/alphagov/email-alert-api/blob/master/lib/reports/subscription_changes_after_switch_to_daily_digest_report.rb#L17