Update `updated_at` when we use `update_all`
update_all doesn't update timestamps when used and so leaves the
'updated_at' timestamp unchanged. This indirectly created a bug as we
use the updated_at field as filter for one of [our reports] which resulted
in nils being passed around.

This updates the other uses of `update_all` in the email system.

[our reports]: https://github.com/alphagov/email-alert-api/blob/master/lib/reports/subscription_changes_after_switch_to_daily_digest_report.rb#L17