Pending ADR for new approach to monitoring and alerting
This is an ADR that tries to succinctly cover the major themes of the
23 page document on "How do we know the Email Alert System is broken"
[1]. It is a pending ADR as we have not yet started the development of
these new alerts which may lead to some of the consequences of the
decision being altered.

There's a few places where it could have appropriate to link to internal
Google Docs (such as incidents or the source document) however it felt
unconventional to link to internal documents in an ADR so I left this
out.

This was a tricky decision to portray as the work behind it was many
decisions. I tried to focus just on the major themes with the
expectation that future implementation documentation will convey more
granular information. In trying to make this succinct there were number
of avenues I chose not to go down, notably:

- Talking about why alerting is embedded in the system and the risks of that
  compared to entirely external alerting (it's hard to get sufficient
  information to call it).
- Aiming to alert about a problem with only a single alert even if it affects
  multiple machines. We're trying to avoid the situation where a problem occurs
  and an engineer gets ~10 alerts about the problem.
- Why it's not appropriate to monitor every content change to the granularity
  as is done in Email Alert monitoring.
- Talking about consistency across GOV.UK alerting approaches. They're mostly
  quite simplistic and there are pretty big gaps where things can break and
  no-one would get called automatically.
- Details as to why the stack boundaries are what they are, why Notify
  monitoring isn't our responsibility for instance.
- How we'd know if the monitoring machines stopped receiving data that powers
  alerts.
- How we may want to adjust retry schedules and attempts to reflect the
  new approaches.

As always in an ADR I ended up in a muddle over tense. I initially tried
to have this in past tense but that didn't flow right given the pending
nature of consequences, so I've instead opted for future tense.

[1]: https://docs.google.com/document/d/1HczXfKL6dJz1OqjJIqJSA8AohUiZJfZbVhsorCJdx0E/edit#