Remove faulty and redundant validation
Previously this was added as part of the introduction of "ended_at"
and the concept of an "active" subscription [1]. The intent was to make
sure that there could be only one active subscription per list/user.
Although the schema-level validation works as expected [2], the way
the model one is written means that all previously ended subscriptions
are considered invalid: the "conditions" option isn't a supported
key, so the validation applies to active *and* inactive subscriptions.
Although we could fix the model validation, the schema-level one makes
it redundant. This removes the faulty / redundant validation.

[1]: https://github.com/alphagov/email-alert-api/pull/474
[2]: https://github.com/alphagov/email-alert-api/blob/11fb84542e6c7f3995f419e4affaf56aa759ec6c/db/schema.rb#L181