Revert "Remove redundant email uniqueness validation"
This reverts commit c118d2c40496eb1e898935062a5dc1dc68d254b7.

Previously we removed this validation on the basis that it was
covered by a DB/schema-level validation. However, it turns out
the specific exception had an impact on the response code, which
changed from a 422 to a 500. This reverts the validations, and
adds a request-level spec to properly document it.