Rewrite content for subscription auth email
This is part of the new design for creating a new subscription. Note
that the new content makes use of the list title, so we need to pass
the SubscriberList, instead of just the "topic_id" (list slug) to the
builder. A consequence is that the endpoint may now 404 when called
with a "topic_id" that does not match a known list, which will just
be relayed by the frontend [1].

[1]: https://github.com/alphagov/email-alert-frontend/blob/236b846b631bb0e2733077c10d0b2d6ebd0842fd/app/controllers/application_controller.rb#L12