Updates our analytics doc
Updates our analytics doc to match our new data [retention ADR].

Removes the information around how to query archived emails in Athena as
this process is now deprecated.

The rake task section has been left alone in this stream of work as
there is a story coming up in the near future to decipher which
analytics tasks are going to be kept/combined. This doc can be updated
once we've made those decisions.

Trello:
https://trello.com/c/tfu9zXuy/534-delete-unused-data-after-a-year

[retention ADR]: https://github.com/alphagov/email-alert-api/blob/master/docs/adr/adr-007-retain-data-for-up-to-one-year.md#decision