Convert failed emails to pending when trying to resend
This task previously relied upon the SendEmailWorker attempting to send
failed emails. This worker no longer supports this and requires emails
to be a pending state.

To resolve this and still provide these support tasks we update these
emails from their "failed" status back to "pending".

This approach, and retrying "failed" emails, isn't ideal as "failed"
implies the sending of an email is done and that we can discard the
email. This method complicates that system and probably should be
rethought or even removed at some point.