Add sidekiq-unique-jobs gem
This gem was [removed from Email Alert API] once before in favour of
database locks. However we now want to re-add and use it for the
SendEmailWorker to prevent duplicate emails being queued up for sending
as this could cause large amounts of pointless build up when the system
is busy. The implementation follows in the coming commits.

[removed from Email Alert API]: https://github.com/alphagov/email-alert-api/pull/1419/commits/7d7b1b9ff43e0eb8567b2f63c90684c03433b506