Use Sidekiq unique jobs for SendEmailWorker
We want to add the Sidekiq unique jobs gem to SendEmailWorker so
duplicate jobs can't be queued up.

The Sidekiq option of `unique_across_queues: true` has been added which
forces the uniqueness constraint to only consider the worker class and job
aprguments (instead of also factoring in a specific queue).

This needed for a few reasons:

* We have different priority queues for sending emails and one email
  could either be on the `send_email_immediate` or the
  `send_email_immediate_high` queue so we should check multiple queues.

* A failing job will be pushed onto the RetrySet queue and this option
  prevents emails being re-queued if they're currently being retried.

If a duplicate job (email) attempts to join a queue then it will be met
with the following uniqueness error:

```
SidekiqUniqueJobs::Client::Middleware
DIG-uniquejobs:1f169d355eae2f3793919f1b5f3e6125 INFO: skipping job with
id (a86359cd2f571d31cfe08ecb) because unique_digest:
(uniquejobs:1f169d355eae2f3793919f1b5f3e6125) already exists
```

There is a concern that a lock is held without the job being in Sidekiq
which would block the future attempt from recovering. This could be
mitigated by a timeout on the lock, however in version 6 of
sidekiq-unique-jobs that doesn't look to be possible. This [may be
possible in version 7].

[may be possible in version 7]: https://github.com/mhenrixon/sidekiq-unique-jobs/blob/master/lib/sidekiq_unique_jobs/lua/lock.lua#L74-L89