Recover old pending emails
This commit adds the recovery of old pending emails.

The fetching of pending emails is done in batches to prevent a large
amount being loaded into memory all at once. A scenario where this might
happen would be if we experienced a high volume of communication failure
with Notify which could lead to a sizeable amount of pending emails
being recovered.