Set Notify template_id as a Rails config variable
This is a step towards removing the extraneous and unconventional Notify
configuration. This config is now handled by the more common
Rails.application.config approach.

We can relatively safely remove the default template_id from the
config/notify.yml file as that is only used outside production for
tests. These tests can be rewritten to avoid the need for it.