Rewrite SendEmailService to not use providers
This rewrite changes how SendEmailService works to use the contents of
the GOVUK_NOTIFY_RECIPIENTS environment variable to decide whether an
email should be sent to Notify or not. If an email should be sent to
Notify the SendEmailService::SendNotifyEmail class is used, otherwise
the SendEmailService::SendPseudoEmail class is used. Both of these
classes have been newly created as replacements for the NotifyProvider
and PseudoProvider.

This change was motivated initially by a problem the overrider in
SendEmailService provided in staging and integration. This utility would
return early when an address wasn't part of an allowed list meaning that
most emails would never appear to leave "pending" status, in those
environments. As well as fixing this problem this acts as a wider
simplification of the configuration mechanisms for sending email.

This has commit introduces a variety of small changes:

- Rather than a configurable provider (by environment variable) there is
  now only sending via Notify or writing sending to a log file
  (Pseudo) - this is all controlled by the GOVUK_NOTIFY_RECIPIENTS env
  var;
- Default Rails configuration means are used to configure Notify rather
  than EmailAlertAPI.config, this is in preparation for this sub-system
  being removed;
- Pseudo emails are no longer written to a dedicated log file, this
  seemed unnecessary, I've also improved the formatting;
- It is no longer possible to use the DelayProvider, should the need for
  this re-arise it could be configured as part of the SendPseudoEmail
  class.