Adapt feature tests for changed SendEmailService
This sets up all request tests to be set to send emails to Notify with
ClimateControl and sets up Notify stubbing. In all our
feature/request tests we seem to be testing as if emails would be sent
through Notify so it makes sense to have this as a default.