Remove EmailAlertAPI.config
This removes the configuration and files involved in the
EmailAlertAPI.config as these files are now redundant given the
application has been migrated to use Rails configuration.

As part of this we can rename the application name to be EmailAlertApi
(as the autoloader expects) to remove the need for special zeitwerk
configuration.